+++ lib/pick_drop_ui/pages/work_page_functionalities/during_navigation.dart
@@ -12,6 +12,8 @@ class during_navigation extends StatefulWidget {
   _during_navigationState createState() => _during_navigationState();
 }
 
+
+
 class _during_navigationState extends State<during_navigation> {
   @override
   Widget build(BuildContext context) {
